Add files via upload
This for stack and queue from SoftUni.

First, you will be given a sequence representing steel. Afterward, you will be given another sequence representing carbon.
You should start from the first steel and try to mix it with the last carbon. If the sum of their values is equal to any of the swords in the table below you should forge the sword corresponding to the value and remove both the steel and the carbon. Otherwise, remove only the steel, increase the value of the carbon by 5 and insert it back into the collection. You need to stop forging when you have no more steel or carbon left.

Input:
40 50 70 120 10 20
30 20 30 20 30 50
Output:

You have forged 3 swords.
Steel left: none
Carbon left: 45, 20, 30
Katana: 2
Shamshir: 1